test(maven): split parser tests by surface
