+++ src/parsers/maven/manifest_test.rs
@@ -9,6 +9,90 @@ use std::fs;
 use std::path::PathBuf;
 use tempfile::TempDir;
 
+fn create_temp_manifest(content: &str) -> (TempDir, PathBuf) {
+    let temp_dir = TempDir::new().expect("Failed to create temp directory");
+    let manifest_path = temp_dir.path().join("MANIFEST.MF");
+
+    fs::write(&manifest_path, content).expect("Failed to write manifest");
+
+    (temp_dir, manifest_path)
+}
+
+#[test]
+fn test_parse_manifest_mf_implementation() {
+    let manifest_path = PathBuf::from("testdata/maven/test2/MANIFEST.MF");
+    let package_data = MavenParser::extract_first_package(&manifest_path);
+
+    assert_eq!(package_data.package_type, Some(PackageType::Maven));
+    assert_eq!(
+        package_data.datasource_id,
+        Some(DatasourceId::JavaJarManifest)
+    );
+    assert_eq!(package_data.name, Some("spring-web".to_string()));
+    assert_eq!(package_data.version, Some("5.3.20".to_string()));
+
+    assert_eq!(package_data.parties.len(), 1);
+    let vendor = &package_data.parties[0];
+    assert_eq!(vendor.r#type, Some("organization".to_string()));
+    assert_eq!(vendor.role, Some("vendor".to_string()));
+    assert_eq!(vendor.name, Some("Spring Framework".to_string()));
+}
+
+#[test]
+fn test_parse_manifest_mf_bundle() {
+    let manifest_path = PathBuf::from("testdata/maven/test3/MANIFEST.MF");
+    let package_data = MavenParser::extract_first_package(&manifest_path);
+
+    assert_eq!(package_data.package_type, Some(PackageType::Osgi));
+    assert_eq!(
+        package_data.datasource_id,
+        Some(DatasourceId::JavaOsgiManifest)
+    );
+    assert_eq!(package_data.name, Some("com.example.mybundle".to_string()));
+    assert_eq!(package_data.version, Some("2.1.0".to_string()));
+
+    assert_eq!(package_data.parties.len(), 1);
+    let vendor = &package_data.parties[0];
+    assert_eq!(vendor.name, Some("Example Corp".to_string()));
+}
+
+#[test]
+fn test_missing_manifest_mf_preserves_manifest_datasource() {
+    let manifest_path = PathBuf::from("/nonexistent/MANIFEST.MF");
+    let package_data = MavenParser::extract_first_package(&manifest_path);
+
+    assert_eq!(package_data.package_type, Some(PackageType::Maven));
+    assert_eq!(
+        package_data.datasource_id,
+        Some(DatasourceId::JavaJarManifest)
+    );
+}
+
+#[test]
+fn test_minimal_manifest_mf_stays_generic_jar() {
+    let content = "Manifest-Version: 1.0\nStart-Class: ${foo.main}\n";
+    let (_temp_dir, manifest_path) = create_temp_manifest(content);
+    let package_data = MavenParser::extract_first_package(&manifest_path);
+
+    assert_eq!(package_data.package_type, Some(PackageType::Jar));
+    assert_eq!(
+        package_data.datasource_id,
+        Some(DatasourceId::JavaJarManifest)
+    );
+    assert_eq!(package_data.name, None);
+    assert_eq!(package_data.version, None);
+    assert_eq!(package_data.purl, None);
+}
+
+#[test]
+fn test_is_match_manifest_mf() {
+    let valid_path = PathBuf::from("/some/path/MANIFEST.MF");
+    let invalid_path = PathBuf::from("/some/path/manifest.mf");
+
+    assert!(MavenParser::is_match(&valid_path));
+    assert!(!MavenParser::is_match(&invalid_path));
+}
+
 #[test]
 fn test_osgi_basic_bundle_detection() {
     let path = PathBuf::from("testdata/osgi/basic/META-INF/MANIFEST.MF");
